ai-service/requirements.txt
# Core AI Service Dependencies
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.5.0
python-dotenv==1.0.0

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.4
numpy==1.26.4

# TTS Models (Coqui TTS)
TTS==0.22.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# Voice Cloning
tortoise-tts==2.8.0

# LLM Integration
groq==0.4.1
requests==2.32.3
httpx==0.25.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# Audio Enhancement
noisereduce==3.0.0
pedalboard==0.8.7

# Music Generation
mido==1.3.0
pretty_midi==0.2.10

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

